Lambdaレイヤーを修正
