+++ lambda/layers/mt_sample_database/requirements.txt
@@ -1,1 +1,2 @@
 sqlalchemy==1.4.29
+psycopg2-binary==2.9.3
